Moved range widgets to one package
