+++ collect_app/src/main/java/org/odk/collect/android/widgets/WidgetFactory.java
@@ -47,6 +47,10 @@ import org.odk.collect.android.widgets.items.SelectOneFromMapWidget;
 import org.odk.collect.android.widgets.items.SelectOneImageMapWidget;
 import org.odk.collect.android.widgets.items.SelectOneMinimalWidget;
 import org.odk.collect.android.widgets.items.SelectOneWidget;
+import org.odk.collect.android.widgets.range.RangeDecimalWidget;
+import org.odk.collect.android.widgets.range.RangeIntegerWidget;
+import org.odk.collect.android.widgets.range.RangePickerDecimalWidget;
+import org.odk.collect.android.widgets.range.RangePickerIntegerWidget;
 import org.odk.collect.android.widgets.utilities.ActivityGeoDataRequester;
 import org.odk.collect.android.widgets.utilities.AudioPlayer;
 import org.odk.collect.android.widgets.utilities.AudioRecorderRecordingStatusHandler;
