+++ collect_app/src/test/java/org/odk/collect/android/widgets/range/RangeIntegerWidgetTest.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.widgets;
+package org.odk.collect.android.widgets.range;
 
 import android.view.MotionEvent;
 import android.view.View;
